move some routes and controllers to admin from auth, change color from yellow to green on some buttons for comparison.
